requirements.txt
# FastAPI & Web Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12

# Google APIs
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Whisper (Speech-to-Text)
faster-whisper==1.0.3
openai-whisper==20240930

# LLM (Gemini)
google-generativeai==0.8.3

# Video Processing
yt-dlp==2024.10.7

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5

# Optional: Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
